+++ Assets/MirrorMarketSDK/Runtime/Services/Interfaces/IMarketplaceService.cs
@@ -26,9 +26,9 @@ namespace MirrorworldSDK.Interfaces
         public void GetNFTDetails(string mintAddress, Action<CommonResponse<SingleNFTResponse>> callBack);
 
         //list
-        public void ListNFT(string mintAddress, decimal price, string confirmation, Action<CommonResponse<ListingResponse>> callBack);
+        public void ListNFT(string mintAddress, float price, string confirmation, Action<CommonResponse<ListingResponse>> callBack);
         
-        public void UpdateNFTListing(string mintAddress, decimal price, string confirmation, Action<CommonResponse<ListingResponse>> callBack);
+        public void UpdateNFTListing(string mintAddress, float price, string confirmation, Action<CommonResponse<ListingResponse>> callBack);
         
         public void CancelNFTListing(string mintAddress, decimal price, string confirmation, Action<CommonResponse<ListingResponse>> callBack);
